requirements.txt
boto3>=1.21.23
Faker>=13.3.2
pyaml>=21.10.1
alive-progress>=2.4.0